tentativas de fazer funcionar mudanças ao vivo de cor e rotação. (o popup ao clicar na forma nao está mais abrindo, tem que corrigir isso)

+++ app/src/main/java/com/example/indoorar/ui/Action.kt
@@ -13,6 +13,6 @@ sealed class Action {
         var end: PointF,
         var selected: Boolean = false,
         var fillColor: Int = "#D9D9D9".toColorInt(),
-        var rotation: Float = 0f
+        var rotation: Int = 0
     ) : Action()
 }
